+++ smt/surrogate_models/qp.py
@@ -11,7 +11,7 @@ import numpy as np
 import scipy
 from smt.surrogate_models.surrogate_model import SurrogateModel
 from smt.utils.caching import cached_operation
-from smt.utils.kriging_utils import standardization
+from smt.utils.misc import standardization
 
 
 class QP(SurrogateModel):
